feature: fixed import problems due to directory rename.
